Moves accounts benches into accounts-db crate
